move init_db and get_sess to database
